requirements.txt

tensorflow>=2.10.0
opencv-python>=4.8.0
numpy>=1.21.0,<1.24.0
pandas>=1.5.0
customtkinter>=5.0.0
Pillow>=9.0.0
pygame>=2.1.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.60.0
